Wire proptests into CI and refresh build tooling
- Register the five proptest suites in Cargo.toml and run them in the
  test workflow with a fixed PROPTEST_SEED
- Bump redis to 1.6, uuid to 1.25, add proptest 1.6
- build.rs falls back to homebrew librtlsdr linking
- Orchestrator gains a process lifecycle module with backend handoff;
  decrypt script and jest config updates
- Drop stale docs assets/performance baselines; onboarding prompt

+++ src/rs/live_stream_test/algorithms.rs
@@ -83,7 +83,7 @@ impl AlgorithmTester {
     &mut self,
     waveform: &[f32],
     timestamp: i64,
-    _center_freq_hz: u32,
+    _center_freq_hz: u64,
     sample_rate_hz: u32,
   ) {
     let peaks = find_peaks(waveform, sample_rate_hz, -60.0, 5);
@@ -152,7 +152,7 @@ impl AlgorithmTester {
     &mut self,
     waveform: &[f32],
     timestamp: i64,
-    _center_freq_hz: u32,
+    _center_freq_hz: u64,
     sample_rate_hz: u32,
   ) {
     if waveform.is_empty() {
@@ -226,7 +226,7 @@ impl AlgorithmTester {
     &mut self,
     iq_bytes: &[u8],
     _timestamp: i64,
-    center_freq_hz: u32,
+    center_freq_hz: u64,
     sample_rate_hz: u32,
   ) {
     println!(
